// src/client/nifidRoutes.ts
export interface JsonSpec {
  path?: string
  name?: string
  type?: string
}

export interface StringObjectParams {
  pattern?: string
  name?: string
}
// ...
export interface SubscriptionParams {
  protocol?: string
  name?: string
  orgID?: string
  brokerHost?: string
  brokerPort?: number
  brokerUsername?: string
  brokerPassword?: string
  brokerCert?: string
  brokerKey?: string
  topic?: string
  dataFormat?: string
  jsonMeasurementKey?: {
    path?: string
    name?: string
    type?: string
  }
  jsonFieldKeys?: JsonSpec[]
  jsonTagKeys?: JsonSpec[]
  jsonTimestamp?: {
    path?: string
    name?: string
    type?: string
  }
  stringMeasurement?: {
    pattern?: string
  }
  stringFields?: StringObjectParams[]
  stringTags?: StringObjectParams[]
  stringTimestamp?: {
    pattern?: string
  }
  bucket?: string
  qos?: number
}
// ...
export interface GetBrokerSubParams {}
// ...
export interface PostBrokerSubParams {
  data: SubscriptionParams
}
// ...
export interface PutBrokerSubParams {
  id: string

  data: SubscriptionParams
}
// ...
export interface DeleteBrokerSubParams {
  id: string
}
// ...
export interface GetBrokerSubStatusParams {
  id: string
}